Fixed user subscription serialization in current user view. Removed some unused API views.
