+++ vdgsa_backend/accounts/views/__init__.py
@@ -1,6 +1,7 @@
 from .change_email import ChangeEmailRequestView as ChangeEmailRequestView
 from .change_email import change_current_user_email_request as change_current_user_email_request
 from .change_email import change_email_confirm as change_email_confirm
+from .current_user_view import CurrentUserView as CurrentUserView
 from .membership_secretary import AllUsersSpreadsheetView as AllUsersSpreadsheetView
 from .membership_secretary import MembershipSecretaryView as MembershipSecretaryView
 from .stripe_checkout import stripe_cancel_view as stripe_cancel_view
@@ -12,4 +13,3 @@ from .subscription import StripeWebhookView as StripeWebhookView
 from .user_profile import UserProfileView as UserProfileView
 from .user_profile import current_user_profile_view as current_user_profile_view
 from .user_registration import UserRegistrationView as UserRegistrationView
-from .views import CurrentUserView as CurrentUserView
